apps: tidied up the code in the weather app
